+++ CHANGES.md
@@ -1,6 +1,6 @@
 ## v0.56.0
  
- * Add careOf (c/o) field to address.
+ * Add additionalAddressInfo field to address.
 
 ## v0.55.0